pyproject.toml
[project]
name = 'gpnmgr'
version = '0.0.1'
description = 'GPN team and access management'
readme = 'README.md'
requires-python = '>=3.13'
authors = [
    { name = 'promasu' },
]
classifiers = [
    'Programming Language :: Python',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3 :: Only',
    'Programming Language :: Python :: 3.13',
    'Programming Language :: Python :: 3.14',
]

dependencies = [
    'Django==6.0',
    'django-auditlog==3.4.1',
    'django-bootstrap5==25.3',
    'django-qr-code==4.2.0',
    'django-localflavor==5.0',
    'gunicorn==23.0.0',
    'psycopg2-binary==2.9.11',
    'Authlib==1.6.6',
    'requests==2.32.5',
    'python-dateutil==2.9.0.post0',
    'ldap3==2.9.1',
]

[project.optional-dependencies]
tests = [
    'coverage==7.11.0',
    'django-stubs==5.2.7',
    'mypy==1.18.2',
    'pylint==4.0.2',
    'pylint-django==2.6.1',
    'types-requests~=2.32.0',
]
docs = [
    'mkdocs==1.6.1',
    'mkdocs-material==9.6.22'
]

[project.urls]
repository = 'https://github.com/entropia/gpnmgr'

[tool.setuptools.packages.find]
where=['src']

[tool.setuptools.package-data]
shiftings = ['py.typed']
